@@ -8,11 +8,9 @@
   },
   "author": "Michael Schoonmaker <michael.r.schoonmaker@gmail.com>",
   "license": "MIT",
-  "peerDependencies": {
-    "react": "~0.11.1"
-  },
   "dependencies": {
     "littlest-dispatcher": "~0.2.3",
+    "react": "~0.12.1",
     "routr": "0.0.1",
     "when": "~3.4.5"
   },
