@@ -8,6 +8,7 @@
  * - Intercepting anchor navigation.
  * - Rehydrating any state received by the server.
  */
+var React = require('react');
 var when = require('when');
 
 /**
@@ -16,10 +17,6 @@ var when = require('when');
  *
  *  - `router`: A valid Router instance. Required.
  *  - `window`: The top-level `window` object. Required.
- *  - `React`: The React library to use. Since React store state internal to
- *      the module instance, sharing the module _itself_ is an invariant.
- *      Though providing this module is recommended, the Client will attempt
- *      to `require` it if not provided.
  *  - `context`: A valid Context instance. If provided, this Context will be
  *      rehydrated with any state provided from upstream Renderers provided a
  *      similar Context. Recommended.
@@ -33,7 +30,6 @@ function Client(options) {
 
   this.router = options.router || null;
   this.window = options.window || (typeof window === 'undefined' ? null : window);
-  this.React = options.React || require('react');
   this.context = options.context || null;
 
   if (!this.router) {
@@ -76,7 +72,7 @@ Client.prototype.renderApplication = function renderApplication(location, perfor
         context: self.context
       };
 
-      self.React.renderComponent(route.body(props), document.body);
+      React.render(route.body(props), document.body);
 
       document.title = self.getTitle(props) || document.title;
     });
