@@ -4,6 +4,7 @@
  */
 var fs = require('fs');
 var util = require('util');
+var React = require('react');
 var when = require('when');
 var whenKeys = require('when/keys');
 
@@ -17,10 +18,6 @@ var whenKeys = require('when/keys');
  *    be loaded from this absolute path. Either `template` or `templatePath` is
  *    required.
  *  - `router`: A valid Router instance. Required.
- *  - `React`: The React library to use. Since React store state internal to
- *      the module instance, sharing the module _itself_ is an invariant.
- *      Though providing this module is recommended, the Client will attempt
- *      to `require` it if not provided.
  *  - `context`: A valid Context instance. If provided, this Context will be
  *      serialized in a way that a similar Context provided to a downstream
  *      Client will be rehydrated with the same state. Recommended.
@@ -34,7 +31,6 @@ function Renderer(options) {
 
   this.template = options.template || null;
   this.router = options.router || null;
-  this.React = options.React || require('react');
   this.context = options.context || null;
 
   if (!this.template && options.templatePath) {
@@ -62,13 +58,13 @@ Renderer.createRenderer = Renderer;
  */
 Renderer.prototype.render = function render(components, context) {
   var html = this.template;
-  var body = this.React.renderComponentToString(components.body);
+  var body = React.renderToString(components.body);
   var head;
 
   body += '<script>window.LITTLEST_ISOMORPH_CONTEXT = ' + JSON.stringify(context || {}) + ';</script>';
 
   if (this._renderHead) {
-    head = this.React.renderComponentToStaticMarkup(components.head);
+    head = React.renderToStaticMarkup(components.head);
     html = html.replace('{head}', head);
   }
 
